api/requirements.txt
uvicorn
fastapi
sqlalchemy
pymysql
mysql-connector-python
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
azure-identity
azure-keyvault-secrets
azure-storage-blob
langchain-community
unstructured
langchain
astrapy>=0.5.3
langchain-community
langchain-openai
langchain-astradb
langchainhub
fastembed
pypdf
langchain-classic
bcrypt==4.0.1